chore: Renaming admin apps
